refactor(neural render): moved utils to the layers package fix(models): imports
