+++ makiflow/models/__init__.py
@@ -21,7 +21,6 @@ from .ssd.detector_classifier import DetectorClassifier
 from .ssd import ssd_utils
 from .ssd.ssd_model import SSDModel
 from .classificator import Classificator
-from .segmentation import Segmentator
 from .rnn.text_recognizer import TextRecognizer
 from .nn_render import NeuralRender
